requirements.txt
python-telegram-bot==5.3.0
django==1.10.5
psycopg2==2.6.2